[SofaPython3] ADD A downcasting implementation.
Implement a downcasting mechanisme to map sofa component to their
corresponding binding.

Without this PR pybind11 is returning the binding corresponding to the lowest possible
class. So if in the c++ code it is written
Base* b;
return py::cast(b)

The python type will be Sofa.Core.Base

If it is
BaseObject* b;
return py::cast(b)

Then the python type will be Sofa.Core.BaseObject

If it nicer if we can get more 'precise' type eg:
m = root.addObject("MechanicalObject")

On solution is to do that explicitly (at minimal cost):
m = Sofa.Component.MechanicalObject( root.addObject("MechanicalObject") )

Or have the downcasting done automatically in the binding (with an higher cost)
m = root.addObject("MechanicalObject")  # return a Sofa.Component.MechanicalObject

NB: Performance issues:
using automatic downcasting do have a performance cost. This is why there is a
caching mechanism. After first downcast to a type the "long" search disapreas.
Ideally we should not use a map for this cache but an hash table and we should use the
BaseClass* getClass() instead of the className. I tried but for a strange reason
the pointer returned by sofa is changing. To me this is a Sofa bug...but once fixed
the overhead of the downcasting will be very low.

Making small benchmark should be nice to validate what I'm saying.