[SofaPython3] Restore the plugins automatic registering of path to python libraries.
When we load a c++ plugin it may be packaged with additional python files
(examples, libraries, or per plugin binding). These files should be visible from the python environment.

This is done by the added code that scans the plugin's directory and register any path matching the following patterns:
pluginName/lib/site-packages
pluginName/lib/python3/site-packages

This behavior was the implemented in the previous version of SofaPython and proved to be useful.