[SofaPython3] Activate 'per component' logging by default when using the Sofa binding
The message pipeline in sofa is rather complex and by default the components does not
stores the messages (info, warning, error) that are emitted. In runSofa the message pipeline
is configured so keep track of eache messages received on a per object basis.
This is done by registering a dedicated message handler in the message pipeline.

In this PR we active this behavior by default in the Sofa python module so the behavior
is consistant with what happens in Sofa.