[bindings/Sofa/tests] Change few tests in ForceField and Node.
The PR is breaking existing code. But the previously working code was,
despite being conveniant was not valid and it was using a "flat" array as
a multi-dimensionnal structure. It was not noticed before because the lists were
converted into a 1D array then into a big string.

Two options were possibles:
a) not allowing this kind of code (using a 1D array to fill a 2D structure)
b) implement kind of de-flattening a python lists into the corresponding multidimmensional matrix

For the simplicity I choose option (a), so I updated the tests to use the 'right' syntax.