[all] FIX the packaging system.
The initial problem is that the SP3_add_python_package (and few others) was using
configure_file() to install python package into the build dir (as well as install dir).

The problem is that configure_file, because of the variable expansion, create a link
between the file and CMake...so each time the file is touched, cmake is re-executed.
To me it make no sense to use configure for every .py unless you really want variable expansion (and thus have an
explicit filename.py.in) for that.

So I refactor the whole thing to:
- stop using configure_file
- make a link between the python files from the source dir to the build dir, so
  (so we can now edit python bug without noticing too late that we touched the files in build directory).
- make an install process that copy the files (and not the links)
- put all the test-files at the same location in various tests