Remove usages of SofaComponentAll
SofaComponentAll was mostly used to load kernel components (SofaBase, SofaSimulation and SofaCore). These are not yet pluginized. The idea with this commit is to remove the dependency of SofaComponentAll which loads much more than the kernel components, and replace it with hard link dependencies until we finish the pluginization of the SOFA kernel.