Fix tests (#65)
* Fix remaining tests

* Remove usages of SofaComponentAll

SofaComponentAll was mostly used to load kernel components (SofaBase, SofaSimulation and SofaCore). These are not yet pluginized. The idea with this commit is to remove the dependency of SofaComponentAll which loads much more than the kernel components, and replace it with hard link dependencies until we finish the pluginization of the SOFA kernel.

* Remove SOFA cache from the CI

SOFA minimal binaries are lightweights are it is taking about the same time to download and install them compared to the caching system. This will simplify CI workflows.

* Use stable SOFA branch v20.12