[SofaPython3] Add an hasObject method in Binding_Node to detect if a node has an object. (#154)
With the current binding it is possible to detect if a node has an object using
the default python hasattr() function. But because of how hasattr() is implement
it imply the generation of the bindinged object regardeless this is usefull or not
from the callee's point of view.

So we add a fast path version, with the corresponding documentation and tests.