[SofaPython3] Call init() by default when creating a python object.
The default behavior should be the most common use case while more
rare scenario should be accessible through explicit call.

Currently the default case is:
   m = r.addObject("MechanicalObject", position=[1,2,3])
   m.init()
   l.rest_position() ## Rest position is created from position at init

This is a bit weird as one may expect that once "created" in python the object is
immediately accessible in an inited state (not doing so imply user to call init() at every lines).

After the PR the default case become:
   m = r.addObject("MechanicalObject", position=[1,2,3])
   l.rest_position()

But for rare use case where it makes sense the no init behavior can be done like that:
   m = r.addObject("MechanicalObject", position=[1,2,3], doNoInit=True)
   l.rest_position() # Rest position is "uninited"
   m.init()          # You can control when you init the object
   l.rest_position() #

Unit tests are provided to validate the behavior.