[SofaPython3] Improve the exception returned when trying to set value from invalid data type.
The existing code to set a data field of type string from string was not correct.
It was converting the python type to string which was leading to allowing things like that:
object.name.value = ["Name"]
print(object.name.value)      # was printing '["Damien"]'

This is not allowed anymore rising an exception returning the type of the objcet passed.
object.name.value = ["Name"]
[ERROR]   [SofaPython3::SceneLoader] RuntimeError: trying to set value for 'name' from a python object of type <class 'list'>