[SofaPython3/Binding] Do not use the binding cache mecanisme when the data is dirty (Fix sofa issue #2761) (#243)
* [SofaPython3/DataHelper.cpp] update data before accessing its bufferInfo

If not done then the generated buffer info can be out-of-sync the data was dirty.
The buffer info is extracting the raw memory address of the contained data. If the data
 array change then the buffer is not valid anymore.

Signed-off-by: Damien Marchal <damien.marchal@univ-lille1.fr>

* [SofaPython3] Refresh the array cache if the data is not valid anymore.

I was convinced that this caching system was gone.
It is not so let's avoid it to provide wrong buffer when the data is not valid anymore.