[SofaPython3] Add a LinkPath object in both the binding and plugin.
The existing syntax to make links in Sofa is the following one:
```python
node.addObject("MeshObjLoader", name="loader")
node.addObject("MechanicalObject", position=node.loader.position.getLinkPath())
```

In addition to be very verbose the getLinkPath() is returning a string (eg: "@/node/loader.position")
Representing linkpath with string has drawback as it force the two objects to be part of the
same simulation graph which is not always the case.

Example of code that does not work as expected:
```python
def MyPrefab(target):
	node = Sofa.Core.Node("MyPrefab")
	node.addObject("MeshObjLoader", name="loader")
	node2.addObject("MechanicalObject", position=target)
	return node

def createScene(root):
	root.addObject("MeshObjLoader", name="loader")
	root.addChild( MyPrefab(loader.position.getLinkPath()) )
```
It does not work because getLinkPath is returning a string, this string is then
used to makes a path query into the scene-graph to locate the object or the data
field to attach to.
But in this case as the object issuing the query is not in the same graph
the query fails.

To solve this issue this PR introduce a LinkPath object which is pointing to a sofa object.
This LinkPath object is exposed in python and can be used to make parenting between objects and data
without the need of graph to path then path to graph conversion (so it is also much faster..
but we don't really care here right ?).

For consistency with the "value" in data field the PR expose this link path
with a dedicated "linkpath" attribute.
So the code is now the following:
```python
root.addObject("MeshObjLoader", name="loader")
root.addChild("MechanicalObject", name="p1", position=root.loader.position.value)      # Copy the value
root.addChild("MechanicalObject", name="p2", position=root.loader.position.linkpath)   # Make a link between data without string query

root.addObject("Mapping", name="loader", input=p1.linkpath)
```