[SofaPyhon3/Plugin] Fix regression introduced in PR #4243
To fix memory error in PR #243

There is the need to get the latest data by calling getValue().

The way it was done in the PR is wrong because it is using a begin/end edit pairs
which trigger the DDG's follower to notify the data is dirty (while it is not).

The consequence was that every call on read access eg:
print(myobjec.position.value)
was triggering the data change mecanisme