[SofaPython3] Separates the pre-loaded python modules out of the python environment (BUGFIX)
The execution of Sofa from a python3 environement is not working anymore because
some features need Sofa & SofaRuntime modules to be loaded but the loading is done when
initializing only in the embeded python interpreter.

To fix this, the PR move the module loading and access into its own singleton that initialize at
first use.

Fixing that is important for 22.06 release.
For the future, refactoring the PythonEnvironment file to split PythonEnvironment in two, to separate what is actually relevant for the PythonEnvironment and what
should be part something like PythonEmbededInterpreter.