[DataHelper] Change behaviour of toBufferInfo for empty vectors
Until now, no difference was made in the toBufferInfo method depending
of the size of the handled data, in particular if this data was an empty
vector (with size = 0). Accessing such a vector through Python code
would called a trigger to the toBufferInfo method, resulting in a call
to the getValuePtr method and an access to the first element (data[0]),
which in this case doesn't exist.

We change this default behaviour by calling the getValuePtr method only
if the vector is not empty, and by returning a null pointer otherwise.