Add proper return value to event overriden in python (#543)
Currently the controllers implemented in python cannot report the the event processing
system in sofa that there is no need anymore to process the event.

The PR allows to do. Every onXXXXXX() that implement an event processor can no return a value.
If it is None or False then processing of event use this to continue the propagation
If it is True (aka: has been processed/no more further), then the SOFA system will not proposed to other component to process the same event